chore(packaging): Rethink now changelog/release messages are maintained.
Previously the GitHub release messages were generated directly from the git
commit log.  Now they are used to generate an entry for a changelog file that
can be edited before it is committed, and the GitHub release message is
extracted from the changelog file.

This new workflow is intended to encourage the writing of better commit
messages, to minimize the effort of maintaining the changelog, and to use
these better changelog entries in the GitHub release messages.